NCBC-1453: Use JSON transcoder when reading updated bucket config
MOTIVATION
----------
It possible to configure a custom document content transcoder where the
transcoder is responsible for encoding and decoding document bodies.
However, if a NotMyVBucket response is received from the server, this
transcoder is used to try and decode the updated bucket config. This is
dangerous because the custom transcoder may not work with JSON and the
response will always be JSON.

MODIFICATIONS
-------------
- Add an overload to IOperation.GetConfig that takes an ITypeTranscoder
  to be used for decoding the response body
- Overload the new method on OperationBase to use the transcoder
  parameter instead of the content transcoder
- Add ServerConfigTranscoder to IClusterController & ClusterController
  that internall resolves to a DefaultTranscoder
- Update all internal usage of GetConfig to use the transcoder in
  IClusterController
- Add unit tests to verify behaviour for both method variants

RESULT
------
When a NotMyVBucket response is received, the SDK will use a JSON
transcoder to decode the response instead of the custom transcoder.

Change-Id: Ic2d0aa63653947070a7b8e18c378ce2b35386d81
Reviewed-on: http://review.couchbase.org/80524
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>