NCBC-1465: Don’t use 127.0.0.1 as default for NodeExt.Hostname
MOTIVATION
----------
As part of NCBC-1457, a default value of 127.0.0.1 was added to
NodeExt.Hostname when being instantiated. However, the NodeAdapater was
checking for a null value and replacing the valid hostname. This has the
side-effect that if a single-node cluster was setup without setting the
hostname during setup, the SDK fails to connect to it and attempts to
connect to 127.0.0.1.

MODIFICATIONS
-------------
- Remove the NodeExt.Hostname default
- Update unit tests to expect Node.Hostname to not have a default value

RESULT
------
NodeExt does not default to 127.0.0.1 and prevent Node.Hostname being
using. This enables the SDK to connect to single-node clusters where the
default hostname has not been changed from 127.0.0.1.

Change-Id: Ib88ac10fb385d9b1bdeb6c529cea2fd38a4cdce2
Reviewed-on: http://review.couchbase.org/80599
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>