NCBC-1463: Add AuthenticationDomain to IUserManager API
MOTIVATION
----------
An update to the UserManagement RFC has added an AuthenticationDomain
parameter that is to be used to modify the http request URI path. An
enumeration of local & external should be added with all API methods
to include
the new parameter.

MODIFICATIONS
-------------
- Add AuthenticationDomain (Local & External)
- Update IUserManager API to take AuthenticationDomain
- Update ClusterManager implementations to take additional parameter and
  to use the domain to form the http request path
- Update unit and integration test suits to provide domain parameter
- Add test case to ensure http request URI is correctly formed based on
  domain

RESULT
------
All of the IUserManager methods now take an AuthenticationDomain
parameter that is used to build the correct path (local vs external).

Change-Id: Icb256a1e5e157d60344faebadc8567ca7b976850
Reviewed-on: http://review.couchbase.org/80946
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>