NCBC-1323: Update CI version and lower build verbosity
MOTIVATION
----------
The build version is set when executing build.ps1, however the original
CI build number is retained even though they can be very different.

The build output is very verbose, we can lower verbosity with a msbuild
switch.

MODIFICATIONS
-------------
- update CI build number to be the same as build.ps1 output
- set msbuild verbosity to quiet

RESULT
------
The output during CI builds is not as verbose.

Change-Id: Idbb08b23023e9ffc134249aba1e872d23f259b64
Reviewed-on: http://review.couchbase.org/81188
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>