NCBC-1206: Allow per-operation timeouts
Motivation
----------
Previously, the SDK allows for operation timeouts to be set globally,
however in certain cases one might want to set the timeout on a
per-operation basis. This commit adds a TimeSpan parameter to most of the
K/V operations that allows the default to be overridden.

Modifications
-------------
 - Add overloads that take a TimeSpan to various K/V operations
 - Add a WithTimeout method to the Sub-Document API
 - Add a GlobalTimeout property to the bucket implementations

Result
------
For most K/V operations a per-operation timeout can be used.

Change-Id: Iba305fdcb23cd9ee75ef43917ede597a4c9d435e
Reviewed-on: http://review.couchbase.org/81264
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>