NCBC-1449: Support RBAC authentication via configuration file
MOTIVATION
----------
With RBAC it is expected that a client provide a single username and
password during bootstrap to authenticate with, however it is not
possible to provide these via configuration. It should be possible to
provide these values in either of .config or .json files.

MODIFICATIONS
-------------
- Add Username and Password properties to ICouchbaseClientDefinition,
  CouchbaseClientDefinition and CouchbaseClientSection
- Check for username and password during client config init and create a
  password authenticator
- Add unit and integration tests to verify behaviour for both
  .config and .json files

RESULT
------
It is now possible to authenticate with a 5.0+ cluster using RBAC by
providing a valid username and password in either .config or .json
files. A username can be provided either directly in the 'username'
property or as part of the server URI (eg connection string).

Change-Id: I0bb76681f00508b39b9b97d47a9ddd9985fdeccc
Reviewed-on: http://review.couchbase.org/81521
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>