NCBC-1472: TTL incorrectly set when value in milliseconds is above ~24 days
Motivation
----------
When a Document.Expiry is set above ~24 days, a casting bug will cause an
overflow creating a negative value to be stored, instead of the expected
positive timestamp.

Modifications
-------------
Update the TimeSpanExtensions.ToTtl(uint) method to create the TimeSpan
using TimeSpan.FromMilliseconds(..) instead of the CTOR.

Result
------
The correct TimeSpan value is created for the expiration of the document.

Change-Id: I1b5c2a0e08b58d6434a34725f5640dcfca363785
Reviewed-on: http://review.couchbase.org/81806
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>