NCBC-1475: Clone does not copy Expiry for some commands
Motivation
----------
For commands that take a expiry (add, replace, etc) in some cases the
expiry is not copied in the Clone method. If an operation is retried, the
Clone method is used and the expiry is not copied from the original
object, so it remains zero; infinite expiration.

Modifications
-------------
- Add Expiry to Clone() method for operations that take an expiry
- Add unit tests for Add, Decrement, Increment, Replace and Set operations
- Add Expiry to Sub-Document operations which take an expiry.
- Add unit tests for Sub-Document operations which take an expiry
- Remove expiration from Increment ctor for consistency

Results
-------
If an operation is retried, Clone will copy the expiry so that it is
properly set for the document.

Change-Id: Ibdce10969118c2927eed9afe60e89126f6cff5fd
Reviewed-on: http://review.couchbase.org/82030
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>