NCBC-1479: Make integration tests spock compatible
MOTIVATION
----------
With server 5.0 RBAC is the default authentication mechanism and it
requires a username and password to provide at the cluster level. The
integration tests need to be able to toggle between previous 5.0
servers and 5.0.

MODIFICATIONS
-------------
- Add SetupEnhancedAuth ICluster extention method that looks to see
  enhancedAuth property is true, and applies a PasswordAuthenticator
  to the cluster
- Add enhancedAuth property to TestSettings
- Add default enhancedAuth value of 'false' to config.json
- Update integration tests to call SetupEnhancedAuth whenever a
  cluster is created to ensure it can correctly authenticate

RESULT
------
Tests that were failing due to RBAC authentication errors are fixed
and can be run.

Change-Id: I17c0404bcc36bc3c59fb8932121db3f672052c31
Reviewed-on: http://review.couchbase.org/82685
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>