NCBC-1491: Don’t dispose of memcached buckets during tests
MOTIVATION
----------
The memcached integration tests dispose of the bucket during each test.
This is not desirable because the connection pool for the bucket will
be destroyed and re-created repeatedly.

MODIFICATIONS
-------------
- create a reference to the memcached bucket and reuse through the tests

RESULT
------
Memcached bucket tests no longer dispose of the bucket for each test.

NOTE: The tests Test_Multi_Upsert and When_Key_Exists_Exists_Returns_True
that continues to use the beer-sample bucket and doesn't used a
memcached bucket.

Change-Id: I5f6f3a395023e09becf88bf33d96e75148f3de9a
Reviewed-on: http://review.couchbase.org/82689
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>