NCBC-1494: Update subdoc XATTR API to match RFC
MOTIVATION
----------
The current implementation of subdoc XATTR support along with other
additional functionality does not follow the RFC. The main issues are
that the subdoc path and document flags are expected to be passed in
unique bytes and that the meaning / value of the flags are mixed
together. Also, the doc flags byte should be omitted if no flags are
set.

MODIFICATIONS
-------------
- Refactor SubdocLookupInFlags and SubdocMutateInFlags to
  SubdocPathFlags and SubdocDocFlags with expected entries and values
- Update IMutateInBuilder, ILookupInBuilder, MutateInBuilder and
  LookupInBuilder to accept the new style flags in place of the old
- Replace the Flags byte on OperationSpec with two flags fields
- Calculate the correct extras length if there are doc flags in
  SubdocSinglularBase
- Set the path and doc flags bytes in SubdocSingularLookupBase and
  SubdocSingluarMutationBase
- Update all unit and integration tests to use the new style flags

RESULT
------
The subdoc XATTR API now matches the RFC and is able to perform all the
expected operations outlines in the spec.

Change-Id: I352b17b5dbf203c04defa2c63d93c74cd45664e7
Reviewed-on: http://review.couchbase.org/82752
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>