NCBC-1498: Make MultiplexIOService work with RBAC
MOTIVATION
----------
MultiplexIOService doesn't work with RBAC because it doesn't perform
the SelectBucket operation during bootstrap. This is because connection
pool is initialized before the IOService is created and this is what
indicates if EnhancedAuthentication is required.

Also, ConnectionPool will created mizSize +1 connections because the
first connection is not counted during initalize.

MODIFICATIONS
-------------
- Ensure all existing connections are both authenticated and
  perform the SelectBucket operations during initialize
- Use _refs.Count instead of _store.Count to work out how many
  additional connections to create
- Execute ConnectionPool.Initalize after IOService has bootstrapped
  to make sure EnhancedAuthentication property is set to indicate
  the connections need to do the SelectBucket operation

RESULT
------
MultiplexIOService correctly executes the SelectBucket operation
before performing operations.

Change-Id: I2b60ad4e46851686794ddfa93d75e28f963c1203
Reviewed-on: http://review.couchbase.org/82766
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>