NCBC-1505: Format exception with Common.Logging and QueryClient
Motivation
----------
Common.Logging calls String.Format in Log.Debug(..). If the incoming
string has any "{" or "}" it will throw a FormatException. This commit
strips the "{" and "}" before calling Log.Debug(..).

Modifications
-------------
Remove "{" and "}" before logging the query request if DEBUG is enabled.

Result
------
FormatException should not be logged or written to stdout.

Change-Id: Icd4ca6b9a19aa75e1b5f1edbae4e7bf38c174fc3
Reviewed-on: http://review.couchbase.org/82835
Reviewed-by: Jaekwon Park <jaekwon.park@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>