NCBC-1502: check if query plan is stale and purge cache if stale
Motivation
----------
If a stale query plan is encountered, the client may fail with the error
"unexpected end of JSON input". This commit purges the query plan cache if
its stale and then retries the request.

Modifications
-------------
 - Check that query plan stale in QueryClient and purge if
   necessary.
 - Add extension method that checks the response to see if its stale
 - Add tests

Result
------
If the query plan is stale, it will be purged from  the cache and the
query retried.

Change-Id: Ife46dc81e58d1f795799df024d456a691ec5c3d8
Reviewed-on: http://review.couchbase.org/82991
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>