NCBC-1436: QueryClient hides base properties HttpClient and DataMapper
Motivation
----------
Both IQueryClient and HttpServiceBase both have DataMapper and HttpClient,
because QueryClient implement or derive from both, these properties are
overriden and hidden causing a compiler warning. This commit removes the
properties from IQueryClient and QueryClient so that the compiler is
removed.

Modifications
-------------
Remove HttpClient and DataMapper properties from IQueryClient and
QueryClient so that the base implementation is used.

Result
------
Compiler warning is removed.

Change-Id: Id4a7fea06677645373fa324f312220db20f11afa
Reviewed-on: http://review.couchbase.org/83159
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>