NCBC-1515: Update AppVeyor AWS credentials
MOTIVATION
----------
The AWS credentials used to upload the zip file to S3 are invalid
and require updating. Also, now we have released 2.5.0 the prefix
for new releases should be updated.

MODIFICATIONS
-------------
- update AWS credentials used to upload zip to S3
- update path prefix for s3 upload

RESULT
------
The AWS credentials used to upload the zip to S3 are now valid and
the path used to upload is correct.

Change-Id: Ic72f05e91448da39fe2860dc0bf74ab32699abf0
Reviewed-on: http://review.couchbase.org/83204
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>