NCBC-1517: Enabled KVErrorMap by default
MOTIVATION
----------
KV Error Map is a new feature in Couchbase Server 5.0, and was
optionally enabled using configuration during development. The default
should now always to be enabled and the tuneable option marked as
obsolete.

MODIFICATIONS
-------------
- Always send XError feature code during Hello in IOServiceBase
- Mark each reference to UseKVErrorMap as obsolete, in
  PoolConfiguration, BucketElement, BucketConfiguration,
  BucketDefinition and ConfigContextBase
- Update tests that set UseKVErrorMap to not set anymore

RESULT
------
The KVErrorMap feature is always requested and the option to disable
this has been marked as obsolete along with tests being updated.

Change-Id: I2914bb27dc27b5d6273937c45d2ab70e632580e8
Reviewed-on: http://review.couchbase.org/83268
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>