NCBC-1510: Create and publish symbols package during builds
MOTIVATION
----------
It is desirable to create symbols nuget packages that can be deployed
alongside full nuget packages to help with debugging. They contains both
pdb files and the original source code and are pushed to dedicated
symbols servers for indexing. Using these symbols packages allows client
applications to debug original source code to help identify issues.

MODIFICATIONS
-------------
- add IncludeSymbols and IncludeSource parameters to msbuild command
- update skip_symbols to false for both nuget and myget in appveyor.yml
- add symbol server url for myget deployments
- remove default nuget feed url for nuget deployments
- add *.nupkg to .gitignore

RESULT
------
An additional nuget package is created that contains both the pdb files
and source code that can be used to help debug a client application. The
extra nuget package is published to either myget or nuget symbols feed
as part of the appveyor deployment step.

Change-Id: I876fb3a98637c7e48fcb15441ca66b42d3ffb111
Reviewed-on: http://review.couchbase.org/83277
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>