NCBC-1521: Remove obsolete properties from http services
MOTIVATION
----------
The IDataMapper and HttpClient properties defined in some of the http
service interfaces and implementation classes are not required. They
were marked as obsolete during the 2.4 development cycle and should be
removed now we are in the 2.5 developnent cycle.

MODIFICATIONS
-------------
- remove IDataMapper and HttpClient property definitions from
  ISerchClient, IQueryClient and IViewClient interfaces
- remove property implementations from SearchClient, QueryClient and
  ViewClientBase

RESULT
------
The obsolete IDataMapper and HttpClient properties have been removed
from each of the http service interfaces and class implementations.

Change-Id: I684e767ed4f169af2a472118fd2cf80f411544d1
Reviewed-on: http://review.couchbase.org/83399
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>