NCBC-1523: FTS - Unexpected character encountered while parsing value
Motivation
----------
If a non-HTTP 4XX error occurs, the server will return an object of
key/value pairs with the pindex and the error message. This fixes a bug
where the client assumed that each error message was a string and not an
object.

Modifications
-------------
 - Change the parsing of the errors property so that it treated each error
   as a key/value.
 - Add unit test and example output file.
 - Fix logic to determine if request was succesful.
 - Rename tests so that its easier to find.

Result
------
If the errors property has values, they will converted into a string and
added to the SearchQueryResult.Errors collection.

Change-Id: I6c7f2409a49033609e37887c2448354381a2752a
Reviewed-on: http://review.couchbase.org/83441
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>