NCBC-1522: Improve exception handling when auth fails
Motivation
----------
When using the ClassicAuthenticator with pre-5.0 clusters, if you do not
call AddBucketCredntials(bucket, password), the client will expectedly
fail to bootstrap, however, the message should indicate that its an
authentication error and not a standard CLR message when an operation is
attempted on an empty list.

Modifications
------------
 - Throw AuthenticationException if the bucket crendentials list is empty.
 - Add unit test

Result
------
If bucket crendentials are not provided, the client will fail-fast and
throw an AuthenticationException.

Change-Id: I0dd01b7fab1b4c5f266c8e8ab5cbc57c900697f2
Reviewed-on: http://review.couchbase.org/83528
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>