NCBC-1528: Don’t run N1QL pretty print tests on unsupported servers
MOTIVATION
----------
There is an integration tests that fails during CI because Couchbase
Server 4.5 does not support the Pretty Print N1QL parameter. The test
should only be run when working with a compatible server.

MODIFICATIONS
-------------
- ignore pretty print tests when server version < 4.6.0

RESULT
------
Pretty print integration tests are only executed when working with a
Couchbase Server version >= 4.6.0.

Change-Id: I3f115e83c56ce3572478fe66b508b04eaf3d7079
Reviewed-on: http://review.couchbase.org/83639
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>