NCBC-1529: Ignore compiler warning for missing XML comments
MOTIVATION
----------
When building the Couchbase library, there are lots of compiler warnings
regarding public properties that do not have XML descriptions. This is
something we would like to improve, but does will be over time.

In the short-term it would be good to ignore these warnings.

MODIFICATIONS
-------------
- add nowarn property with value 1591 to Couchbase.csporj

RESULT
------
There are no more compiler warnings for missing XML comments on public
properties when building the Couchbase library.

Change-Id: I9b76dca6aa5440e2547abb67f18b1a641d551816
Reviewed-on: http://review.couchbase.org/83642
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>