NCBC-1524: Add missing license information
MOTIVATION
----------
All files in the Couchbase SDK should have appropriate license
information. This is done using a LICENSE file in the root of the
repository and each file should have a region at the end with
license details.

MODIFICATIONS
-------------
- add LICENSE file to root of repository with license information
- add missing license region to all files in Couchbase project

RESULT
------
License information is now available both through a LICENSE file in the
root of the repository and in each file in the Couchbase project.

Change-Id: I271da98ef2d5491ed2e12f547397833a075c9e8f
Reviewed-on: http://review.couchbase.org/83959
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>