NCBC-1545: Add BucketNotConnected response code
MOTIVATION
----------
It has been found that during rebalance of adding new nodes the .NET SDK
does not fail authentication when the SelectBucket operation returns a
non-success. This is not captured and the connection is then used for
subsequent opeations which fail with BucketNotConnected response code.

MODIFICATIONS
-------------
- add ResponseCode.BucketNotConnected
- set ResponseCode.BucketNotConnected as non-retryable in
  OperationResult.ShouldRetry
- set BucketNotConnected as a status that does not need any further
  processing in OperationResult.SetException
- update both MemcachedConfigContext and CouchbaseConfigContext to
  perform authentication in say steps as during CCCP bootstrap
- add long line to ConnectionPoolBase.EnableEnhancedAuthentication to
  indicate when SelectBucke operations fail
- ensure Authenticate is called for connections in
  SharedConnectionPool.Aquire

RESULT
------
The correct authentication process should be followed both during
bootstrap and when new bucket configurations are received for both
Memcached and Couchbase buckets.

Change-Id: I9f29a7035f3fc237f00896a3139db5c73555c401
Reviewed-on: http://review.couchbase.org/84059
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>