NCBC-1551: Fix expiration bug with sync bulk upsert
Motivation
----------
NCBC-1206 introduced per-operation timeouts for each operation.
Unfortunately this change incorrectly interpreted the operation timeout
as the TTL instead for the (obsolete) synchronous bulk upsert operation.

This commit fixes this bug and ensures that the timeout is honored and
expiration is not incorrectly set.

Modifications
-------------
- Explicitly set expiration to 0 in all synchronous bulk upserts.
- Pass this expiration as well as the timeout to the underlying Upsert
method.

Result
------
The expiration is now correctly set to 0 for all synchronous bulk
upserts.

Change-Id: I95f07471e4a869fd1314e8898f973695a121734e
Reviewed-on: http://review.couchbase.org/84226
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>