NCBC-1553: Add unit tests to root Couchbase namespace
MOTIVATION
----------
The test coverage for a number of important classes in the root
Couchbase namespace is either 0 or a low value. Test coverage should be
improved to both prove expected behaviour and prevent bugs being
introduced to the code base.

MODIFICATIONS
-------------
- Add unit test coverage for MemcachedBucket with regards to unsupported
  durability overloads
- Add unit test coverage for ClusterInfo, Cluster, CouchbaseBucket,
  OperationResult and ResponseExtensionsTests

RESULT
------
The test coverage on classes in the root namespace has improved with
more code paths being exercised by both unit and integration tests.

Change-Id: I44e12c58f27e54c58e2c8fcca1bcc6eff1df5980
Reviewed-on: http://review.couchbase.org/84426
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>