NCBC-1559: Invalid default timeout is set for Memcached Upsert
Motivation
----------
MemcachedBucket.Upsert(key, value) has a bug where the expiration is set
to the GlobalTimeout value and instead of the key having an infinite
timeout, it times out nearly immediately causing the key to be evicted
from the bucket.
Other overloads and operations do not appear to be impacted; the work
around is to use any of the other overloads.

Modifications
-------------
Changed default value for expiration to be TimeSpan.Zero which will be in
an infinite expiration for the given key.

Results
-------
Fixes bug where the GlobalTimeout (2500ms) for the operation is used
instead of the expiration of TimeSpan.Infinite. Keys inserted with this
method will have infinite lifetimes.

Change-Id: I4ff798e4201976172f68155b662b6da02ff8d3a2
Reviewed-on: http://review.couchbase.org/84502
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>