NCBC-1569: Mark IBucket.Upsert with dictionary of keys obsolete
MOTIVATION
----------
All IBucket.Upsert methods that take a dictionary of keys of objects
should be marked as obsolete because they are inefficient and internally
use Parallel.ForEach which can cause issues. Instead the preferred
method is to create and wait for UpsertAsync tasks to complete.

MODIFICATIONS
-------------
- mark IBucket.Upsert methods that take a dictionary of keys and objects
  with timeouts as obsolete

RESULT
------
All IBucket.Upsert methods that use a dictionary of keys and objects are
marked as obsolete with a recommendation to use the async version
instead.

Change-Id: Ibe3c2dafcfa8dba752a7d73407ec748222636729
Reviewed-on: http://review.couchbase.org/85056
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>