NCBC-1570: Implement Memcached.Upsert for multiple items including timeout
MOTIVATION
----------
Memcached.Upsert with a dictionary of keys and objects, ParallelOptions
and a TimeSpan should be implemented and not return
NotSupportedException. This also affects the Upsert overload that does
not provide a timeout because it uses the above overload internally.

MODIFICATIONS
-------------
- add implementation for MemcachedBucket.Upsert that takes a dictionary
  of keys and objects, ParallelOptions and a TimeSpan

RESULT
------
Memcached.Upsert when passing a dictionary of keys and objects,
ParallelOptions and a TimeSpan is now implemented and does not return a
NotSupportedException.

Change-Id: I45adf69627e69761ecac31ac1556e3a868b2f5e5
Reviewed-on: http://review.couchbase.org/85059
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>