NCBC-1565: Return NotSupportedException for Exists and ExistsAsync in MemcachedBucket
MOTIVATION
----------
Exists and ExistsAsync implementations in MemcachedBucekt try to send
Observer operations to cluster nodes. However, these are not supported
operations for Memcached buckets and instead should fast-fail with a
NotSupportedException.

MODIFICATIONS
-------------
- return NotSupportedException for Exists and ExistsAsync methods in
  MemcachedBucekt

RESULT
------
Both Exists and ExistsAsync in MemcachedBucket return NotSupported
instead of trying to send Observer operations to the server.

Change-Id: I116cb47f9c1aebc955b89fe90bee8a384bb60969
Reviewed-on: http://review.couchbase.org/85060
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>