NCBC-1567: Add IBucket.ReplaceAsync with durability requirements and timeout
MOTIVATION
----------
There is a missing variant of ReplaceAsync on the IBucket interface, and
implementations in both MemcachedBucket and CouchbaseBucket, that takes
a list of documents, durability requirements and timeout that should be
added.

MODIFICATIONS
-------------
- add ReplaceAsync with a list of documents, PersistTo,
  ReplicatTo and TimeSpan to IBucket
- add implementation to MemcachedBucket that throws
  NotSupportedException
- add CouchbaseBucket implementation that executes tasks as expected
  using all the provided parameters
- add integration test to verify Couchbase Bucket behaviour

RESULT
------
There is now a IBucket.ReplaceAsync variant that can take a list of
documents, PersistTo, ReplicateTo and a timeout. Where MemcachedBucket
returns a NotSupportedException and CouchbaseBucket has a real
implementation.

NOTE: Unit test to verify Memcached Bucket throws NotSupportedException
is part of NCBC-1553.

Change-Id: Ia798de9770b7fab2299a6d7f67c0f256f8034db7
Reviewed-on: http://review.couchbase.org/85064
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>