NCBC-1592: BucketCache (non-static API for accessing buckets)
Motivation
----------
BucketCache provides a non-static API for accessing buckets (just like the
static API of ClusterHelper), which will be cached for subsequent requests.
BucketCache can be used with an IoC container (registered as a singleton
instance) which makes the referencing code testable and makes the IoC
container responsible for teardown (by implementing IDisposable).

Modifications
-------------
 - add IBucketCache interface
 - add BucketCache implementing IBucketCache
 - add BucketCache unit tests

Result
------
The static API of ClusterHelper is maintained and a non-static API through
BucketCache is provided for use with an IoC container.