NCBC-1579: Use ConnectionPool.Aquire in PooledIOService constructors
MOTIVATION
----------
PooledIOService requires a connection during construction to check
server features. However, the connection is retrieved using
Connections.FirstOrDefault which can return a null connection and is not
null-checked. Instead, ConnectionPool.Aquire should be used to get the
connection as it will manage creation of the connection if required.

MODIFICATIONS
-------------
- use ConnectionPool.Aquire instead of
  ConnectionPool.Connections.FirstOrDefault

RESULT
------
PooledIOService retrieves a connection from the ConnectionPool using
Aquire instead of using the LINQ query FirstOrDefault.

Change-Id: I7939159cad238c1106961c39bc407fa07b31a500
Reviewed-on: http://review.couchbase.org/85452
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>