NCBC-1580: Fix failing unit tests
MOTIVATION
----------
After merging NCBC-1565 and NCBC-1579 there are some unit and
integration test failures. This is because Aquire is now called
twice during the PooledIOService unit tests and Exists should
throw a NotSupportedException with a MemcachedBucket.

MODIFICATIONS
-------------
- expect Aquire to be called twice during PooledIOService tests
- remove the Exists test from MemcachedBucketTests

RESULT
------
The unit and integration tests pass as expected.

NOTE: The unit tests for verifying NotSupportedException is thrown for
Exists and ExistsAsync for MemcachedBuckets is contained in NCBC-1553.

Change-Id: I7689bf7c12036f3d6012579f20f2516d1da2d79b
Reviewed-on: http://review.couchbase.org/85523
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>