NCBC-1584: Update ketama hash generation to match RFC sample
MOTIVATION
----------
The code that generates a ketama hash ring is not clear and does not
match the example in the RFC. For simplicity and clarity they should
match.

MODIFICATIONS
-------------
- update KetamaMapper.Initialize to use be clearer and match example in
  RFC

RESULT
------
The code for generating the hash ring is clearer and matches the example
in the RFC.

NOTE: The existing unit test KetamaKeyMapperTests.Test_Hash_Ring verifies
the hash ring continues to be correctly constructed.

Change-Id: Idd627ae4550aa438ec468d1c652f060bf637bc6e
Reviewed-on: http://review.couchbase.org/85688
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>