NCBC-1575: Add info to CarrierPublication to use forceSasl for LDAP
MOTIVATION
----------
SCRAM is not supported when using LDAP user authentication and an
exception is caught and logged. It would be beneficial to include some
additional instructions in the log to tell the user to set the
forceSaslPlain setting to true if using LDAP.

MODIFICATIONS
-------------
- extend log message when catching AuthenticationException in
  CarrierPublicationProvider to check set forceSaslPlain if using LDAP

RESULT
------
When an AuthenticationException is caught during CCCP bootstrap, the log
indicates that if using LDAP the forceSaslPlain setting should be set to
true.

Change-Id: I9c80489d645a77ff3b5eb31360e854bdd6211ec2
Reviewed-on: http://review.couchbase.org/85838
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>