NCBC-1600: Correctly parse hostname during bootstrap
MOTIVATION
----------
Bootstrapping is failing when connecting to a Couchbase Cluster using
IPv4 addresses. This is because the NodeExt can have a null or empty
value so the Node.Hostname is used, however this value can include the
port which needs to be removed.

MODIFICATIONS
-------------
- parse the hostname and strip port if present in host, IPv4 and IPv6
  addresses
- add unit tests to verify hostname is correctly parsed for a mixture of
  host, IPv4 and IPv6 including both with and without ports

RESULT
------
When creating a NodeAdapter the hostname is correctly parsed and strips
the port from both IPv4 and IPv6 addresses.

Change-Id: Iee01a0a9cc33aeab632644cc70c02d89ecc43249
Reviewed-on: http://review.couchbase.org/86325
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>