NCBC-1555: DecodeString method doesn't manage null string
Motivation
----------
The DefaultTranscoder would use an empty string when instead it should use
null when a value is null for a key.

Modifications
-------------
 - Change DecodeString to return null instead of an empty string when the
   body of the response is empty.

 - Add DefaultTranscoderTests to Couchbase.UnitTests

Result
------
If a key is stored without a value for a string, the SDK will respect this
and not return an empty string; null will be returned instead.

Change-Id: I48c59c1d8266e1af539b84f6b1564e0a018c1397
Reviewed-on: http://review.couchbase.org/86345
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>