NCBC-1608: Log query context ID when writing query timings
MOTIVATION
----------
QueryTimer logs timing information for a particular query, however the
query context ID is not included which can make it difficult to
correlate with other log entries.

MODIFICATIONS
-------------
- store and include the query context ID when writing query timings to
  the log
- move QueryTimer tests from Couchbase.Tests to Couchbase.UnitTests
- copy Couchbase.Tests.Fakes.FakeLog to Couchbase.UnitTests

RESULT
------
When writing query timings to the log, the query context ID is included
to help correlate with other log messages.

Change-Id: Ib8d08b212285a107fbf421b5437c594a5f7c88cd
Reviewed-on: http://review.couchbase.org/86714
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>