NCBC-1593: Read mutation token for non-generic operations
MOTIVATION
----------
When executing a Delete operation which is based on OperationBase and
mutation tokens are enabled, the result has an empty mutation token
which causes the operation to fail durability checks.

MODIFICATIONS
-------------
- try to read the response body using ReadExtras virtual method when
  processing responses in OperationBase
- add unit tests for both Remove and RemoveAsync to verify behaviour

RESULT
------
Operations that inherit OperationBase will try to read extras when
processing the result which includes reading the mutation token.

Change-Id: I0630bb7802ed70616b196a6a1bf5cc918e706b34
Reviewed-on: http://review.couchbase.org/87095
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>