NCBC-1602: Implement Certificate Authentication for KV
Motivation
----------
Allows for x509 certificates to be used for authenticating the client with the
server. Only works for K/V; Views, Analytics, FTS and N1QL are not currently supported.
Requires that X509 certificates are generated and installed on the server and
a client .pfx be generated from the public key and is installed on the
application server.

Modifications
-------------
 - Updated SslConnection so that if EnableCertificateAuthentication is
   enabled and a CertifcateFactory exists, the client will use them to
   authenticate.
 - By pass authentication in ConnectionPoolBase if
   EnableCertificateAuthentication.
 - Added the EnableCerificateAuthentcation flag to BucketConfiguration and
   related configuration classes.
 - Added CertifcateFactory that provides generic implementatations
   of Func<X509CertifcateCollection2> for fetching x509 certs from disk
   and cert store.
 - Added CertificateFactoryTests.cs for testing the CertificateFactory.

Result
------
The client can authenticate with the server using X509 certificates if
they are properly configured. Only K/V is supported - Views and N1QL does
support certificate authentication. See
https://developer.couchbase.com/documentation/server/current/security/security-x509certsintro.html
for details.

Change-Id: I8e4b48a2d98c90586ae70f9b7e3084141da0bb81
Reviewed-on: http://review.couchbase.org/87131
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>